+++ docs/docs/.vitepress/utils/sidebar.ts
@@ -40,9 +40,13 @@ export const sidebar = {
         { text: 'Text 文本', link: '/components/text' },
         { text: 'Layout 布局', link: '/components/layout' },
         { text: 'Space 间距', link: '/components/space' },
-        { text: 'Color 色彩', link: '/components/color' },
-        { text: 'Interface 接口类型', link: '/components/interface' }
+        { text: 'Color 色彩', link: '/components/color' }
       ]
+    },
+    {
+      text: '类型接口',
+      link: '/components/',
+      items: [{ text: 'Interface 接口类型', link: '/components/interface' }]
     },
     {
       text: '配置组件',