+++ packages/fighting-design/_interface/index.d.ts
@@ -1,2 +1,3 @@
 export * from './auto'
 export * from './event'
+export * from './lang'